Apply final changes from code review
